Add files via upload
In this code, we have a function for each calculation. Therefore, each function will get the initial file and functions aren't linked together. My problem for ordering in a alphabetic form isn't still solved. I appreciate any help.
Here is the problem.
In this project, you have to write a program that reads the scores of different people from a csv file and performs the following calculations on the scores and saves the resulting values in a file.


1- calculate the average of each person and save it along with the name of each person, the output order of the names must be exactly equal to the order of the input file.

2- Save the rates in ascending order along with the name of each person.

3- Save the top three GPAs with each person's name.

4- Save the three low GPAs without each person's name.

5- Calculate and store the average grades.